fix: mark placeholder tests and rename helper module (#1468)
- Add @pytest.mark.skip to 5 placeholder tests in test_host_filtering.py
- Rename tests/issues/517/test_helpers.py to helpers.py
- Update imports in 3 E2E test files to use helpers instead of test_helpers
