+++ tests/integration/test_host_filtering.py
@@ -58,61 +58,55 @@ class TestAPIHostnameFiltering:
             # Hostnames should not be placeholders
             assert not host.startswith("<") and not host.endswith(">")
 
+    @pytest.mark.skip(reason="Requires upload endpoint setup and authentication")
     def test_upload_invalid_hostname_sanitized(self, client, db):
         """
         Test that uploading data with invalid hostname results in empty string in database.
 
         This verifies the entry-point sanitization logic.
         """
-        # Note: This test would require actual upload endpoint setup
-        # and authentication, which may not be available in test environment
-        # The logic is tested in unit tests, this is for integration verification
         pass
 
 
 class TestDatabaseHostnameFiltering:
     """Test database-level hostname filtering."""
 
+    @pytest.mark.skip(reason="Requires database setup with test data")
     def test_sql_filter_condition(self, db):
         """
         Test that SQL WHERE clause correctly filters invalid hostnames.
 
         This verifies the SQL layer filtering works correctly.
         """
-        # Note: This test would require database setup with test data
-        # The logic is tested in unit tests, this is for integration verification
         pass
 
+    @pytest.mark.skip(reason="Requires database setup with test data")
     def test_summary_service_double_filter(self, summary_service, db):
         """
         Test that summary service applies both SQL and Python filtering.
 
         This verifies the double-filtering strategy works correctly.
         """
-        # Note: This test would require database setup with test data
-        # The logic is tested in unit tests, this is for integration verification
         pass
 
 
 class TestLogAuditing:
     """Test logging for audit trail."""
 
+    @pytest.mark.skip(reason="Requires log capture setup")
     def test_invalid_hostname_logged_on_upload(self):
         """
         Test that invalid hostnames are logged when sanitized.
 
         This verifies the audit trail requirement.
         """
-        # Note: This test would require log capture setup
-        # The logic is tested in unit tests, this is for integration verification
         pass
 
+    @pytest.mark.skip(reason="Requires log capture setup")
     def test_sql_filter_miss_logged_as_warning(self, summary_service, db):
         """
         Test that SQL filter misses are logged as warnings.
 
         This verifies the defensive logging requirement.
         """
-        # Note: This test would require log capture setup
-        # The logic is tested in unit tests, this is for integration verification
         pass
