+++ tests/issues/517/e2e_codex_real_session.py
@@ -28,7 +28,7 @@ import sys
 import time
 
 import requests
-from test_helpers import (
+from helpers import (
     BASE_URL,
     PROJECT_ROOT,
     REMOTE_TEST_HOST,
