+++ tests/issues/517/helpers.py
